Enable buffers in the queue from client to background reader
This changes the channel used in the client and minion to be
buffered, which allows for a faster Post().

During the course of working on this, it has been revealed that
the reader goroutine was dependent on the synchronous processing
of the queue. If we use a buffered channel, the writer goroutine
must wait for the reader goroutine to finish processing its queue
first, otherwise pending messages in the queue are never processed.
There are a few changes that for those, as well.